CurvilinearCalculus.jl

CurvilinearCalculus.GenericCoordinates

CurvilinearCalculus.isorthogonal

CurvilinearCalculus.GenericCoordinates — Type

GenericCoordinates(r::CoordinateMapping, q::SVector{3,Coordinate})

Defines a (non-)orthogonal coordinate system with fields

q Coordinates, g_cov Covariant basis vectors (in Cartesian coordinates), e_cov Covariant unit vectors (in Cartesian coordinates), g_contra Contravariant basis vectors (in Cartesian coordinates), e_contra Contravariant unit vectors (in Cartesian coordinates), F transformation matrix new coordinates to Cartesian, G Metric (defined by dot products of covariant basis vectors), invG Inverse of metric, J Volume element, g J^2 or det(G), Γ Christoffel symbols Γⁱⱼₖ → Γ[i,j,k]

CurvilinearCalculus.isorthogonal — Method

Check if coordinate system is orthogonal